Commit last bit of code from the Peachtree Regional (though not used)
This was in effort to make a quick interface for being able to change
the autonomous configuration without having to keep rebuilding the code
and deploy it on the robot. Of course there is some bug in this code
somewhere so it was never used (at least yet).

This also was trying to get it so that we could disable the voltage
ramp when trying to stop the tote elevator but that didn't want to work
either.